Display existing hard links in result sets
This is a breaking change in the plain text output,
but somehow the list displayed in case of existing hard links
was arbitrary, since the all the hardlinked filenames were not displayed.


Here's a sample JSON result with this patch:

    {
      "file_size": 9216,
      "paths": [
        "test_tree/f09-1_5.raw",
        "test_tree/f09-4_5.raw"
      ],
      "links": {
        "test_tree/f09-1_5.raw": [
          "test_tree/f09-2_5.raw",
          "test_tree/f09-3_5.raw"
        ],
        "test_tree/f09-4_5.raw": [
          "test_tree/f09-5_5.raw"
        ]
      }
    }

Here the 5 files have the same contents, but there are two hardlink groups:
    "test_tree/f09-1_5.raw"
    "test_tree/f09-2_5.raw"
    "test_tree/f09-3_5.raw"
are hard-linked, and
    "test_tree/f09-4_5.raw"
    "test_tree/f09-5_5.raw"
are hard-linked.


Here's the same set displayed With the regular text output:

    Group #5 (2 files * 9216 bytes):
    test_tree/f09-1_5.raw
     test_tree/f09-2_5.raw
     test_tree/f09-3_5.raw
    test_tree/f09-4_5.raw
     test_tree/f09-5_5.raw

(The link file names are indented using 1 space character.)